Update earthquakes.md - Orr Chapel Quake
Epicenter: 8397 West Sheepneck Road Sandy Hook, Maury county, Tennessee USA

The Creator is shaking the ground where Orr Chapel once stood, to draw attention to the CIA's murder of all of its identified members.

A total of 82 people were killed initially after being identified as members of the church.

Orr Chapel was a black church, as were the original inhabitants of the unincorporated city of Sandy Hook, TN. The Crawley family and the Rogers family.

The Crawley family lived along Crawley Hill Road and the Rogers family along West Sheepneck Road, both families lived there from around 1888 until 2022. The families owned the land free and clear. My mother's mother was a Crawley and her father was a Rogers, so I represent both families.

The surrounding Maury county's black community has been greatly impacted. The first wave of losses ripped through the community as the CIA felt the need to eliminate anyone associated with the members of Orr Chapel and their families, or the Crawley and Rogers' families of Sandy Hook. In more recent waves, the death toll is comprised of individuals and families that were associated with the earlier waves or those who have publicly questioned or commented on the number of missing from the black community.